croc_sentinel_systems/api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
paho-mqtt==2.1.0
bcrypt==4.2.1
PyJWT==2.10.1
cryptography==44.0.2
